Backend_app/requirements.txt


fastapi
langchain
langgraph
arxiv
langchain-groq
langchain-community
pymongo
uvicorn
python-jose
passlib
bcrypt==4.0.1

python-multipart
pypdf
arxiv
#pymupdf
#pydantic==1.10.2

#Aws lambda deployment requirements
mangum

#Local tests requirements
#langgraph-cli[inmem] #Needed for langgraph studio debugging
#




